+++ generators/config/templates/drivers/webkit.rb
@@ -1,0 +1,6 @@
+# :webkit driver
+
+CapybaraHelpers.load_driver_gem!(:webkit, 'capybara-webkit', 'capybara-webkit')
+Capybara::Webkit.configure do |config|
+  config.allow_url(Howitzer.app_host)
+end