+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Colorized output added
 - New option **user_agent**, allows custom user agent setup 
 - **mailgun_sleep_time** deprecated, **mail_sleep_time** used instead for all mail adapters
-
+- Capybara drivers configuration were split to separate files
 
 ### Bugfixes
 - **window_maximized** option fixed for chrome browser on MacOS